DEV: started tests revision

+++ spectrochempy/core/processors/autosub.py
@@ -70,7 +70,7 @@ def autosub(dataset, ref, *ranges, dim='x', method='vardiff', return_coefs=False
     >>> ref = A[0, :]  # let's subtrack the first row
     >>> B = A.autosub(ref, [3900., 3700.], [1600., 1500.], inplace=False)
     >>> B
-    NDDataset: [float32]  a.u. (shape: (y:55, x:5549))
+    NDDataset: [float64]  a.u. (shape: (y:55, x:5549))
     """
 
     # output dataset
